chore: add quick update
