+++ apps/admin/src/app/post/[postId]/_feature/types.ts
@@ -8,4 +8,6 @@ export type PostContextType = {
   onFileExplorerClose: () => void;
   onMediaBrowse: () => void;
   settings?: Setting;
+  saving: boolean;
+  setSaving: (saving: boolean) => void;
 };
